src/main/webapp/app/entities/batches/delete/batches-delete-dialog.component.ts
import { Component } from '@angular/core';
import { FormsModule } from '@angular/forms';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import SharedModule from 'app/shared/shared.module';
import { IBatches } from '../batches.model';
import { BatchesService } from '../service/batches.service';
import { ITEM_DELETED_EVENT } from 'app/config/navigation.constants';

@Component({
  standalone: true,
  templateUrl: './batches-delete-dialog.component.html',
  imports: [SharedModule, FormsModule],
})
export class BatchesDeleteDialogComponent {
  batches?: IBatches;

  constructor(protected batchesService: BatchesService, protected activeModal: NgbActiveModal) {}

  cancel(): void {
    this.activeModal.dismiss();
  }

  confirmDelete(id: string): void {
    this.batchesService.delete(id).subscribe(() => {
      this.activeModal.close(ITEM_DELETED_EVENT);
    });
  }
}
